# openapi.yaml
openapi: 3.0.1
info:
  title: TodoApi
  version: '1.0'
paths:
  /api/TodoItems:
    get:
      tags:
        - TodoItems
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItem'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItem'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItem'
    post:
      tags:
        - TodoItems
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoItem'
          text/json:
            schema:
              $ref: '#/components/schemas/TodoItem'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TodoItem'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/TodoItem'
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
            text/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
  '/api/TodoItems/{id}':
    get:
      tags:
        - TodoItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/TodoItem'
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
            text/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
    put:
      tags:
        - TodoItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoItem'
          text/json:
            schema:
              $ref: '#/components/schemas/TodoItem'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TodoItem'
      responses:
        '200':
          description: Success
    delete:
      tags:
        - TodoItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
components:
  schemas:
    TodoItem:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        isComplete:
          type: boolean
      additionalProperties: false